works up to heroku deployment

+++ funs_support.py
@@ -1,3 +1,4 @@
+import os
 import io
 import requests
 import contextlib
@@ -11,9 +12,12 @@ def capture(fun,arg):
     return output
 
 # Save download
-def download(url, path=None):
-    req = requests.get(url)
+def download(url, path=None, overwrite=False):
     if path is None:
-        path = req.url.split('/')[-1]
-    with open(path, 'wb') as file:
-        file.write(req.content)
+        path = url.split('/')[-1]
+    if (not os.path.exists(path)) | overwrite:
+        req = requests.get(url)
+        with open(path, 'wb') as file:
+            file.write(req.content)
+    else:
+        print('File will not downloaded')
